+++ frontends/php/include/discovery.inc.php
@@ -1,7 +1,7 @@
 <?php
 /*
-** Zabbix
-** Copyright (C) 2000-2011 Zabbix SIA
+** ZABBIX
+** Copyright (C) 2000-2005 SIA Zabbix
 **
 ** This program is free software; you can redistribute it and/or modify
 ** it under the terms of the GNU General Public License as published by
@@ -19,13 +19,15 @@
 **/
 ?>
 <?php
-	require_once('include/perm.inc.php');
 
+require_once('include/perm.inc.php');
 
+?>
+<?php
 	function check_right_on_discovery($permission){
 		global $USER_DETAILS;
 
-		if($USER_DETAILS['type'] >= USER_TYPE_ZABBIX_ADMIN){
+		if( $USER_DETAILS['type'] >= USER_TYPE_ZABBIX_ADMIN ){
 			if(count(get_accessible_nodes_by_user($USER_DETAILS, $permission, PERM_RES_IDS_ARRAY)))
 				return true;
 		}
@@ -33,24 +35,23 @@
 	}
 
 	function svc_default_port($type_int){
-		$typePort = array(
-			SVC_SSH =>		'22',
-			SVC_LDAP =>		'389',
-			SVC_SMTP =>		'25',
-			SVC_FTP =>		'21',
-			SVC_HTTP =>		'80',
-			SVC_POP =>		'110',
-			SVC_NNTP =>		'119',
-			SVC_IMAP =>		'143',
-			SVC_AGENT =>	'10050',
-			SVC_SNMPv1 =>	'161',
-			SVC_SNMPv2c =>	'161',
-			SVC_SNMPv3 =>	'161',
-			SVC_HTTPS =>	'443',
-			SVC_TELNET =>	'23',
-		);
+		$port = 0;
 
-		return isset($typePort[$type_int]) ? $typePort[$type_int] : 0;
+		switch($type_int){
+			case SVC_SSH: $port = '22'; break;
+			case SVC_LDAP: $port = '389'; break;
+			case SVC_SMTP: $port = '25'; break;
+			case SVC_FTP: $port = '21'; break;
+			case SVC_HTTP: $port = '80'; break;
+			case SVC_POP: $port = '110'; break;
+			case SVC_NNTP: $port = '119'; break;
+			case SVC_IMAP: $port = '143'; break;
+			case SVC_AGENT: $port = '10050'; break;
+			case SVC_SNMPv1: $port = '161'; break;
+			case SVC_SNMPv2: $port = '161'; break;
+			case SVC_SNMPv3: $port = '161'; break;
+		}
+		return $port;
 	}
 
 	function discovery_check_type2str($type=null){
@@ -66,11 +67,9 @@
 			SVC_TCP => S_TCP,
 			SVC_AGENT => S_ZABBIX_AGENT,
 			SVC_SNMPv1 => S_SNMPV1_AGENT,
-			SVC_SNMPv2c => S_SNMPV2_AGENT,
+			SVC_SNMPv2 => S_SNMPV2_AGENT,
 			SVC_SNMPv3 => S_SNMPV3_AGENT,
 			SVC_ICMPPING => S_ICMPPING,
-			SVC_TELNET => _('Telnet'),
-			SVC_HTTPS => _('HTTPS'),
 		);
 
 		if(is_null($type)){
@@ -80,21 +79,19 @@
 		else if(isset($discovery_types[$type]))
 			return $discovery_types[$type];
 		else
-			return false;
+			return S_UNKNOWN;
 	}
 
-	function discovery_check2str($type, $key, $port){
+	function discovery_check2str($type, $snmp_community, $key_, $port){
 		$external_param = '';
 
-		if(!empty($key)){
-			switch($type){
-				case SVC_SNMPv1:
-				case SVC_SNMPv2c:
-				case SVC_SNMPv3:
-				case SVC_AGENT:
-					$external_param = ' "'.$key.'"';
-					break;
-			}
+		switch($type){
+			case SVC_SNMPv1:
+			case SVC_SNMPv2:
+			case SVC_SNMPv3:
+			case SVC_AGENT:
+				$external_param = ' "'.$key_.'"';
+				break;
 		}
 		$result = discovery_check_type2str($type);
 		if((svc_default_port($type) != $port) || ($type == SVC_TCP))
@@ -114,53 +111,182 @@
 	return '';
 	}
 
-	function discovery_status2str($status=null){
-		$discovery_statuses = array(
-			DRULE_STATUS_ACTIVE => _('Active'),
-			DRULE_STATUS_DISABLED => _('Disabled'),
-		);
-
-		if(is_null($status)){
-			return $discovery_statuses;
+	function discovery_status2str($status_int){
+		switch($status_int){
+			case DRULE_STATUS_ACTIVE:	$status = S_ACTIVE;		break;
+			case DRULE_STATUS_DISABLED:	$status = S_DISABLED;		break;
+			default:
+				$status = S_UNKNOWN;		break;
 		}
-		else if(isset($discovery_statuses[$status]))
-			return $discovery_statuses[$status];
-		else
-			return _('Unknown');
+
+	return $status;
 	}
 
 	function discovery_status2style($status){
 		switch($status){
-			case DRULE_STATUS_ACTIVE: $status = 'off'; break;
-			case DRULE_STATUS_DISABLED: $status = 'on'; break;
-			default: $status = 'unknown'; break;
+			case DRULE_STATUS_ACTIVE:	$status = 'off';	break;
+			case DRULE_STATUS_DISABLED:	$status = 'on';		break;
+			default:
+				$status = 'unknown';	break;
 		}
-		return $status;
+
+	return $status;
 	}
 
-	function discovery_object_status2str($status=null){
-		$statuses = array(
-			DOBJECT_STATUS_UP => S_UP,
-			DOBJECT_STATUS_DOWN => S_DOWN,
-			DOBJECT_STATUS_DISCOVER => S_DISCOVERED,
-			DOBJECT_STATUS_LOST => S_LOST,
-		);
+	function discovery_object_status2str($status){
+		$str_stat[DOBJECT_STATUS_UP] = S_UP;
+		$str_stat[DOBJECT_STATUS_DOWN] = S_DOWN;
+		$str_stat[DOBJECT_STATUS_DISCOVER] = S_DISCOVERED;
+		$str_stat[DOBJECT_STATUS_LOST] = S_LOST;
 
-		if(is_null($status)){
-			order_result($statuses);
-			return $statuses;
-		}
-		else if(isset($statuses[$status]))
-			return $statuses[$status];
-		else
-			return S_UNKNOWN;
+		if(isset($str_stat[$status]))
+			return $str_stat[$status];
+
+		return S_UNKNOWN;
 	}
 
 	function get_discovery_rule_by_druleid($druleid){
 		return DBfetch(DBselect('select * from drules where druleid='.$druleid));
 	}
 
+	function set_discovery_rule_status($druleid, $status){
+		return DBexecute('update drules set status='.$status.' where druleid='.$druleid);
+	}
+
+	function add_discovery_check($druleid, $type, $ports, $key, $snmp_community,
+			$snmpv3_securityname, $snmpv3_securitylevel, $snmpv3_authpassphrase, $snmpv3_privpassphrase)
+	{
+		// no need to store those items in DB if they will not be used
+		if($snmpv3_securitylevel == ITEM_SNMPV3_SECURITYLEVEL_NOAUTHNOPRIV){
+			$snmpv3_authpassphrase = '';
+			$snmpv3_privpassphrase = '';
+		}
+		if($snmpv3_securitylevel == ITEM_SNMPV3_SECURITYLEVEL_AUTHNOPRIV){
+			$snmpv3_privpassphrase = '';
+		}
+
+		$dcheckid = get_dbid('dchecks', 'dcheckid');
+		$result = DBexecute('insert into dchecks (dcheckid,druleid,type,ports,key_,snmp_community'.
+				',snmpv3_securityname,snmpv3_securitylevel,snmpv3_authpassphrase,snmpv3_privpassphrase) '.
+				' values ('.$dcheckid.','.$druleid.','.$type.','.zbx_dbstr($ports).','.
+				zbx_dbstr($key).','.zbx_dbstr($snmp_community).','.zbx_dbstr($snmpv3_securityname).','.
+				$snmpv3_securitylevel.','.zbx_dbstr($snmpv3_authpassphrase).','.zbx_dbstr($snmpv3_privpassphrase).')');
+
+		if(!$result)
+			return $result;
+
+		return $dcheckid;
+	}
+
+	function add_discovery_rule($proxy_hostid, $name, $iprange, $delay, $status, $dchecks, $uniqueness_criteria){
+		if(!validate_ip_range($iprange)){
+			error(S_INCORRECT_IP_RANGE);
+			return false;
+		}
+
+		// checking to the duplicate of the name
+		if(CDRule::exists(array('name' => $name))){
+			error(S_DISCOVERY_RULE.SPACE.'['.$name.']'.SPACE.S_ALREADY_EXISTS_SMALL);
+			return false;
+		}
+
+		$druleid = get_dbid('drules', 'druleid');
+		$result = DBexecute('insert into drules (druleid,proxy_hostid,name,iprange,delay,status) '.
+			' values ('.$druleid.','.$proxy_hostid.','.zbx_dbstr($name).','.zbx_dbstr($iprange).','.$delay.','.$status.')');
+
+		if($result && isset($dchecks)){
+			$unique_dcheckid = 0;
+			foreach($dchecks as $id => $data){
+				$data['dcheckid'] = add_discovery_check($druleid, $data['type'], $data['ports'], $data['key'],
+						$data['snmp_community'], $data['snmpv3_securityname'], $data['snmpv3_securitylevel'],
+						$data['snmpv3_authpassphrase'], $data['snmpv3_privpassphrase']);
+				if ($uniqueness_criteria == $id && $data['dcheckid'])
+					$unique_dcheckid = $data['dcheckid'];
+			}
+			DBexecute('UPDATE drules'.
+					' SET unique_dcheckid='.$unique_dcheckid.
+					' WHERE druleid='.$druleid);
+
+			$result = $druleid;
+		}
+
+		return $result;
+	}
+
+	function update_discovery_rule($druleid, $proxy_hostid, $name, $iprange, $delay, $status, $dchecks,	$uniqueness_criteria, $dchecks_deleted){
+		if( !validate_ip_range($iprange) ){
+			error(S_INCORRECT_IP_RANGE);
+			return false;
+
+		}
+
+		// checking to the duplicate of the name
+		$drule = get_discovery_rule_by_druleid($druleid);
+		if(strcmp($drule['name'], $name)){
+			if(CDRule::exists(array('name' => $name))){
+				error(S_DISCOVERY_RULE.SPACE.'['.$name.']'.SPACE.S_ALREADY_EXISTS_SMALL);
+				return false;
+			}
+		}
+
+		$result = DBexecute('update drules set proxy_hostid='.$proxy_hostid.',name='.zbx_dbstr($name).',iprange='.zbx_dbstr($iprange).','.
+			'delay='.$delay.',status='.$status.' where druleid='.$druleid);
+
+		if($result && isset($dchecks)){
+			$unique_dcheckid = 0;
+			foreach($dchecks as $id => $data){
+				if(!isset($data['dcheckid'])){
+					$data['dcheckid'] = add_discovery_check($druleid, $data['type'], $data['ports'], $data['key'],
+							$data['snmp_community'], $data['snmpv3_securityname'], $data['snmpv3_securitylevel'],
+							$data['snmpv3_authpassphrase'], $data['snmpv3_privpassphrase']);
+				}
+				if ($uniqueness_criteria == $id && $data['dcheckid'])
+					$unique_dcheckid = $data['dcheckid'];
+			}
+
+			DBexecute('UPDATE drules'.
+					' SET unique_dcheckid='.$unique_dcheckid.
+					' WHERE druleid='.$druleid);
+		}
+
+		if($result && isset($dchecks_deleted) && !empty($dchecks_deleted))
+			delete_discovery_check($dchecks_deleted);
+
+	return $result;
+	}
+
+	function delete_discovery_check($dcheckids){
+		$actionids = array();
+// conditions
+		$sql = 'SELECT DISTINCT actionid '.
+				' FROM conditions '.
+				' WHERE conditiontype='.CONDITION_TYPE_DCHECK.
+					' AND '.DBcondition('value', $dcheckids, false, true);	// FIXED[POSIBLE value type violation]!!!
+
+		$db_actions = DBselect($sql);
+		while($db_action = DBfetch($db_actions))
+			$actionids[] = $db_action['actionid'];
+
+// disabling actions with deleted conditions
+		if (!empty($actionids)){
+			DBexecute('UPDATE actions '.
+					' SET status='.ACTION_STATUS_DISABLED.
+					' WHERE '.DBcondition('actionid', $actionids));
+
+// delete action conditions
+			DBexecute('DELETE FROM conditions '.
+					' WHERE conditiontype='.CONDITION_TYPE_DCHECK.
+					' AND '.DBcondition('value', $dcheckids, false, true));	// FIXED[POSIBLE value type violation]!!!
+		}
+
+		DBexecute('DELETE FROM dservices WHERE '.DBcondition('dcheckid', $dcheckids));
+
+		DBexecute('DELETE FROM dchecks WHERE '.DBcondition('dcheckid', $dcheckids));
+	}
+
 	function delete_discovery_rule($druleid){
+		$result = true;
+
 		$actionids = array();
 // conditions
 		$sql = 'SELECT DISTINCT actionid '.
@@ -184,7 +310,23 @@
 					" AND value='$druleid'");
 		}
 
-		$result = DBexecute('delete from drules where druleid='.$druleid);
+		if($result){
+			$db_dhosts = DBselect('select dhostid from dhosts'.
+					' where druleid='.$druleid.' and '.DBin_node('dhostid'));
+
+			while ($result && ($db_dhost = DBfetch($db_dhosts)))
+				$result = DBexecute('delete from dservices where'.
+						' dhostid='.$db_dhost['dhostid']);
+		}
+
+		if ($result)
+			$result = DBexecute('delete from dhosts where druleid='.$druleid);
+
+		if ($result)
+			$result = DBexecute('delete from dchecks where druleid='.$druleid);
+
+		if ($result)
+			$result = DBexecute('delete from drules where druleid='.$druleid);
 
 		return $result;
 	}
