+++ src/zabbix_server/housekeeper/housekeeper.c
@@ -1,6 +1,6 @@
 /*
-** Zabbix
-** Copyright (C) 2000-2011 Zabbix SIA
+** ZABBIX
+** Copyright (C) 2000-2005 SIA Zabbix
 **
 ** This program is free software; you can redistribute it and/or modify
 ** it under the terms of the GNU General Public License as published by
@@ -19,7 +19,6 @@
 
 #include "common.h"
 #include "db.h"
-#include "dbcache.h"
 #include "log.h"
 #include "daemon.h"
 #include "zbxself.h"
@@ -33,6 +32,8 @@ extern unsigned char	process_type;
  * Function: housekeeping_process_log                                         *
  *                                                                            *
  * Purpose: process table 'housekeeper' and remove data if required           *
+ *                                                                            *
+ * Parameters:                                                                *
  *                                                                            *
  * Return value: SUCCEED - information removed successfully                   *
  *               FAIL - otherwise                                             *
@@ -178,43 +179,79 @@ static int	housekeeping_sessions(int now)
 	return SUCCEED;
 }
 
-static void	housekeeping_alerts(int now)
+static int	housekeeping_alerts(int now)
 {
 	const char	*__function_name = "housekeeping_alerts";
-	int		deleted, alert_history;
+	int		alert_history;
+	DB_RESULT	result;
+	DB_ROW		row;
+	int		res = SUCCEED;
+	int		deleted;
 
 	zabbix_log(LOG_LEVEL_DEBUG, "In %s() now:%d", __function_name, now);
 
-	deleted = DBexecute("delete from alerts where clock<%d",
-			now - *(int *)DCconfig_get_config_data(&alert_history, CONFIG_ALERT_HISTORY) * SEC_PER_DAY);
-	zabbix_log(LOG_LEVEL_DEBUG, "deleted %d records from table 'alerts'", deleted);
+	result = DBselect("select alert_history from config");
 
-	zabbix_log(LOG_LEVEL_DEBUG, "End of %s()", __function_name);
+	if (NULL == (row = DBfetch(result)) || SUCCEED == DBis_null(row[0]))
+	{
+		zabbix_log(LOG_LEVEL_ERR, "no records in table 'config'");
+		res = FAIL;
+	}
+	else
+	{
+		alert_history = atoi(row[0]);
+
+		deleted = DBexecute("delete from alerts where clock<%d", now - alert_history * SEC_PER_DAY);
+
+		zabbix_log(LOG_LEVEL_DEBUG, "deleted %d records from table 'alerts'", deleted);
+	}
+	DBfree_result(result);
+
+	zabbix_log(LOG_LEVEL_DEBUG, "End of %s():%s", __function_name, zbx_result_string(res));
+
+	return res;
 }
 
-static void	housekeeping_events(int now)
+static int	housekeeping_events(int now)
 {
 	const char	*__function_name = "housekeeping_events";
 	int		event_history;
 	DB_RESULT	result;
-	DB_ROW		row;
+	DB_RESULT	result2;
+	DB_ROW		row1;
+	DB_ROW		row2;
 	zbx_uint64_t	eventid;
+	int		res = SUCCEED;
 
 	zabbix_log(LOG_LEVEL_DEBUG, "In %s() now:%d", __function_name, now);
 
-	result = DBselect("select eventid from events where clock<%d",
-			now - *(int *)DCconfig_get_config_data(&event_history, CONFIG_EVENT_HISTORY) * SEC_PER_DAY);
+	result = DBselect("select event_history from config");
 
-	while (NULL != (row = DBfetch(result)))
+	if (NULL == (row1 = DBfetch(result)) || SUCCEED == DBis_null(row1[0]))
 	{
-		ZBX_STR2UINT64(eventid, row[0]);
+		zabbix_log(LOG_LEVEL_ERR, "no records in table 'config'");
+		res = FAIL;
+	}
+	else
+	{
+		event_history = atoi(row1[0]);
 
-		DBexecute("delete from acknowledges where eventid=" ZBX_FS_UI64, eventid);
-		DBexecute("delete from events where eventid=" ZBX_FS_UI64, eventid);
+		result2 = DBselect("select eventid from events where clock<%d", now - event_history * SEC_PER_DAY);
+
+		while (NULL != (row2 = DBfetch(result2)))
+		{
+			ZBX_STR2UINT64(eventid, row2[0]);
+
+			DBexecute("delete from acknowledges where eventid=" ZBX_FS_UI64, eventid);
+			DBexecute("delete from events where eventid=" ZBX_FS_UI64, eventid);
+		}
+		DBfree_result(result2);
 	}
 	DBfree_result(result);
 
-	zabbix_log(LOG_LEVEL_DEBUG, "End of %s()", __function_name);
+	zabbix_log(LOG_LEVEL_DEBUG, "End of %s():%s", __function_name, zbx_result_string(res));
+
+	return res;
 }
 
 /******************************************************************************
@@ -348,6 +385,10 @@ void	main_housekeeper_loop()
 
 /* Transaction is not required here. It causes timeouts under MySQL. */
 /*		DBcommit();*/
+
+/*		zbx_setproctitle("housekeeper [vacuuming database]");*/
+
+/*		DBvacuum();*/
 
 		DBclose();
